+++ src/autosar/xml/writer.py
@@ -390,6 +390,7 @@ class Writer(_XMLWriter):
             'ExclusiveAreaNestingOrder': self._write_exclusive_area_nesting_order,
             'ExecutableEntityActivationReason': self._write_executable_entity_activation_reason,
             'ExclusiveAreaRefConditional': self._write_exclusive_area_ref_conditional,
+            'SwcExclusiveAreaPolicy': self._write_swc_exclusive_area_policy,
             'RunnableEntityArgument': self._write_runnable_entity_argument,
             'RunnableEntity': self._write_runnable_entity,
             'AsynchronousServerCallReturnsEvent': self._write_async_server_call_returns_event,
@@ -4668,6 +4669,29 @@ class Writer(_XMLWriter):
                 self._write_exclusive_area_ref(elem.exclusive_area, "EXCLUSIVE-AREA-REF")
             self._leave_child()
 
+    def _write_swc_exclusive_area_policy(self, elem: ar_element.SwcExclusiveAreaPolicy) -> None:
+        """
+        Writes complex type AR:SWC-EXCLUSIVE-AREA-POLICY
+        Multi-tagged: False
+        """
+        assert isinstance(elem, ar_element.SwcExclusiveAreaPolicy)
+        tag = "SWC-EXCLUSIVE-AREA-POLICY"
+        if elem.is_empty:
+            self._add_content(tag)
+        else:
+            self._add_child(tag)
+            self._write_swc_exclusive_area_policy_group(elem)
+            self._leave_child()
+
+    def _write_swc_exclusive_area_policy_group(self, elem: ar_element.SwcExclusiveAreaPolicy) -> None:
+        """
+        Writes group AR:SWC-EXCLUSIVE-AREA-POLICY
+        """
+        if elem.api_principle is not None:
+            self._add_content("API-PRINCIPLE", ar_enum.enum_to_xml(elem.api_principle))
+        if elem.exclusive_area is not None:
+            self._write_exclusive_area_ref(elem.exclusive_area, "EXCLUSIVE-AREA-REF")
+
     def _write_abstract_access_point(self, elem: ar_element.AbstractAccessPoint) -> None:
         """
         Writes group AR:ABSTRACT-ACCESS-POINT
@@ -5492,10 +5516,20 @@ class Writer(_XMLWriter):
 
         Most of it will be implemented in a future version
         """
+        if elem.ar_typed_per_instance_memory:
+            self._add_child("AR-TYPED-PER-INSTANCE-MEMORYS")
+            for item in elem.ar_typed_per_instance_memory:
+                self._write_variable_data_prototype(item, "VARIABLE-DATA-PROTOTYPE")
+            self._leave_child()
         if elem.events:
             self._add_child("EVENTS")
             for event in elem.events:
                 self._write_rte_event_element(event)
+            self._leave_child()
+        if elem.exclusive_area_policies:
+            self._add_child("EXCLUSIVE-AREA-POLICYS")
+            for policy in elem.exclusive_area_policies:
+                self._write_swc_exclusive_area_policy(policy)
             self._leave_child()
         if elem.port_api_options:
             self._add_child("PORT-API-OPTIONS")